pyproject.toml
[tool.poetry]
name = "hybrid-search"
version = "0.1.0"
description = ""
authors = ["Bui Thu Ha <thuhak22ba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
langchain = "^0.2.2"
rank-bm25 = "^0.2.2"
singlestoredb = "^1.3.1"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.3"
sentence-transformers = "^3.0.0"
pypdf = "^4.2.0"
langchain-huggingface = "^0.0.2"
openai = "^1.32.0"
fastapi = "^0.111.0"
gradio = "^4.36.0"
cohere = "5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
